ROB now hides items past its length (#1913)
* fixed bug in node TryGet

* ROB hides sliced-off end of the linked list

* Factored out the slow path in Http Parser

* Renamed interfaces

* Update Cli, CoreFx, CoreFx, CoreFx, CoreSetup, CoreSetup to preview1-007522, preview1-25907-04, preview1-25907-04, preview1-25907-04, preview1-25907-02, preview1-25907-02, respectively (#1910)

* Fixed PR review issues

+++ src/System.Buffers.Experimental/System/Range.cs
@@ -4,15 +4,15 @@
 namespace System {
     public struct Range
     {
-        public Range(int index, int length)
+        public Range(long index, long length)
         {
             Index = index;
             Length = length;
         }
-        public readonly int Index;
-        public readonly int Length;
+        public readonly long Index;
+        public readonly long Length;
 
-        public void Deconstruct(out int index, out int length)
+        public void Deconstruct(out long index, out long length)
         {
             index = Index;
             length = Length;
